backend_server/requirements.txt
Django~=5.2.2
djangorestframework~=3.16.0
django-cors-headers==4.7.0
pillow==11.2.1
psycopg2-binary==2.9.10
python-decouple==3.8
djangorestframework_simplejwt==5.5.0